% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elexonDynamic.R
\name{elexonDynamic}
\alias{elexonDynamic}
\title{elexonDynamic}
\usage{
elexonDynamic(key, from = Sys.Date() - 60, to = Sys.Date() - 59,
  test = TRUE)
}
\arguments{
\item{key}{Your personal scripting key from elexon. Find out more at https://www.elexonportal.co.uk}

\item{from}{This is the start date/datetime of the dataset}

\item{to}{This is the end date/datetime of the dataset}

\item{test}{This is set to TRUE by default. Set this argument to FALSE if you want to use the live version of the API.}
}
\description{
This function pulls dynamic data from Elexon/BMRS, given a user's key, start and end dates.
    The data returned are given in a list due to their varying column lengths, separated by each item.
}
\examples{
\dontrun{
elexonDynamic(
key = "948ghmgpe",
from = "2018-01-01",
to = "2019-01-01",
test = TRUE)
}


}
