% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elexonList.R
\name{elexonList}
\alias{elexonList}
\title{elexonList}
\usage{
elexonList()
}
\description{
This simply gives a list of all the datasets you can pull from Elexon/BMRS
}
\examples{
\dontrun{
elexonList()}
\dontrun{
View(elexonList())
}



}
