% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elexonPhysical.R
\name{elexonPhysical}
\alias{elexonPhysical}
\title{elexonPhysical}
\usage{
elexonPhysical(key, from = Sys.Date() - 20, to = Sys.Date() - 19,
  test = TRUE)
}
\arguments{
\item{key}{Your personal scripting key from elexon. Find out more at https://www.elexonportal.co.uk}

\item{from}{This is the start date/datetime of the dataset}

\item{to}{This is the end date/datetime of the dataset}

\item{test}{This is set to TRUE by default. Set this argument to FALSE if you want to use the live version of the API.}
}
\description{
This function pulls physical data from Elexon/BMRS, given a user's key, start and end dates.
    All half-hourly periods are used, accounting for daylight savings as well. The test parameter is set to TRUE by default, as datasets ranging more than a mere week will result in pullinh very large datasets.
}
\examples{
\dontrun{
elexonPhysical(
key = "948ghmgpe",
from = "2018-01-01",
to = "2019-01-01",
test = TRUE)
}


}
