% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapi.R
\docType{data}
\name{listAPI}
\alias{listAPI}
\title{A list of all the dataset names available via BMRS and their parameters
 @docType data}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 95 rows and 10 columns.}
\usage{
listAPI
}
\description{
A list of all the dataset names available via BMRS and their parameters
 @docType data
}
\keyword{datasets}
