% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elexonData.R
\name{elexonData}
\alias{elexonData}
\title{elexonData}
\usage{
elexonData(dataset, key, from, to)
}
\arguments{
\item{dataset}{The dataset you are pulling from BMRS/Elexon.}

\item{key}{Your personal scripting key from elexon. Find out more at https://www.elexonportal.co.uk}

\item{from}{This is the start date/datetime of the dataset}

\item{to}{This is the end date/datetime of the dataset}
}
\description{
This function pulls data from Elexon/BMRS, given a user's key, start and end dates and the dataset in question
    Please note: it does not matter if BMRS requires the dates to be in a different format to "yyyy-mm-dd".
    The Relexon package will take care of this. Just enter the dates in the usual format!
}
\examples{
\dontrun{
elexonData(dataset = "ROLSYSDEM", key = "bne89n4f", from = "2018-01-01", to = "2019-01-01")
}


}
